requirements.txt
gradio==4.44.0
pypdf==4.3.1
python-docx==1.1.2
langdetect==1.0.9
plotly==5.22.0
networkx==3.3
Pillow==10.4.0
python-dotenv==1.0.1
google-generativeai==0.7.2
pydub==0.25.1
audioop-lts==0.2.1
pytesseract==0.3.10
sentence-transformers>=2.2.0
Whoosh>=2.7.0
scikit-learn>=1.3.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart>=0.0.9
chromadb
passlib[bcrypt]
python-jose[cryptography]